fix centering when using threeparttable
When threeparttable=T and a caption is set the table is not being centered on page. In order for the table to be centered on the page, \centering needs to be before \begin{threeparttable}.  Right now, \centering is being put after \caption and the deleted lines insert \begin{threeparttable} between the two.  By removing those lines, \being{threeparttable} is inserted after centering, but before \begin{tabular}.  This results in the table being centered on page.