Data frame input as alternative to named vector
Allow a two column data frame as input instead of a named vector. There are issues with unicode characters in the names of named vectors and these issues are side stepped if the input is a data frame with a character vector with the header names and a numeric vector with the column spans.